Alteração para novo formato de rotas
