+++ lib/login_money.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+
+class LoginMoney extends StatelessWidget {
+  const LoginMoney({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Container();
+  }
+}
