+++ lib/login_tinder.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+
+class LoginTinder extends StatelessWidget {
+  const LoginTinder({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Container();
+  }
+}
