+++ lib/main.dart
@@ -1,5 +1,9 @@
 import 'package:flutter/material.dart';
 
+import 'login_money.dart';
+import 'login_tinder.dart';
+import 'menu.dart';
+
 void main() {
   runApp(MyApp());
 }
@@ -17,60 +21,5 @@ class MyApp extends StatelessWidget {
         '/tinder': (context) => LoginTinder()
       },
     );
-  }
-}
-
-class Menu extends StatelessWidget {
-  const Menu({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-        child: MaterialApp(
-      home: Scaffold(
-        appBar: AppBar(
-          title: Text('Menu - Leitura de Mockups'),
-        ),
-        body: Center(
-          child: Column(
-            mainAxisAlignment: MainAxisAlignment.center,
-            children: [
-              SizedBox(height: 8.0),
-              ElevatedButton(
-                onPressed: () {
-                  Navigator.pushNamed(context, '/login_money');
-                },
-                child: Text("LOGIN MONEY"),
-              ),
-              SizedBox(height: 8.0),
-              ElevatedButton(
-                onPressed: () {
-                  Navigator.pushNamed(context, '/tinder');
-                },
-                child: Text("LOGIN TINDER"),
-              ),
-            ],
-          ),
-        ),
-      ),
-    ));
-  }
-}
-
-class LoginMoney extends StatelessWidget {
-  const LoginMoney({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Container();
-  }
-}
-
-class LoginTinder extends StatelessWidget {
-  const LoginTinder({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Container();
   }
 }
